+++ feature/fido2/src/androidMain/kotlin/com/chimali/fido2/domain/service/Fido2Service.kt
@@ -19,8 +19,6 @@ interface Fido2Service {
         userDisplayName: String,
     ): Outcome<CredentialId, DomainError>
 
-    suspend fun authenticateWithCredential(rpId: RpId): Outcome<CredentialId, DomainError>
-
     fun getAllCredentials(): Flow<PasskeyCredential>
 
     suspend fun deleteCredential(credentialId: CredentialId): Outcome<Unit, DomainError>